Create README.md
This is a Gobang game software developed using Visual C#, which realizes the human-machine
battle, and the stand-alone version of Human-human battle. Regardless of the game mode, user can
regret the game, redo, save the game and read the game during the chess process. The manmachine
battle mode uses an algorithm that only considers one step, so that the computer has the
most basic artificial intelligence, but it needs to be further improved. In this paper, firstly, it
gives a specific introduction to the functionality and UI of the game, then reflects on objectoriented
programming and changes, and gives the class diagram of the entire program and the
timing diagram of some moments. Finally, the pattern used in the design and development
process is summarized, including the structure of the program, the structure of the class, and the
operating mode of the pattern.